fix errors related to byte/string conversion
There were a lot of errors like "a bytes-like object is required,
not 'str'" and the opposite error.

https://stackoverflow.com/questions/33054527/python-3-5-typeerror-a-bytes-like-object-is-required-not-str-when-writing-t
points out that if you open a file in binary mode, you need to write
bytes, not text.  It turns out that Popen with universal_newlines
will open the pipe in text mode rather than binary mode; whereas the
regular file opening calls here are already in text mode, so converting
to utf-8 is not required when writing strings (I hope that means
Python already does that by default).