Fix blocklist; less verbose
Now only does the block list after parsing the XML.

There is scope to do this better - consider the case
of (1) A adds spam, (2) B adds innocent text, and
then (3) C removes spam. By skipping (1), this means
commit revision (2) will be saved with the innocent
text AND the spam from user A but listed as from B.

i.e. The current crude user block skip only really
works if spam was removed as the next revision to
a page.